+++ factura_electronica/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
-__version__ = '3.0.0'
+__version__ = '4.0.0'
