+++ factura_electronica/resources_facelec/facelec_generator.py
@@ -300,7 +300,7 @@ def construir_xml(serie_original_factura, nombre_del_cliente, prefijo_serie, ser
                 item_factura_json['cantidad'] = float(sales_invoice_item[i]['qty'])
                 item_factura_json['codigoProducto'] = str(sales_invoice_item[i]['item_code'])
                 item_factura_json['descripcionProducto'] = str((sales_invoice_item[i]['item_name']))
-                item_factura_json['detalleImpuestoIva'] = float((sales_invoice_item[i]['facelec_sales_tax_for_this_row']))
+                item_factura_json['detalleImpuestosIva'] = float((sales_invoice_item[i]['facelec_sales_tax_for_this_row']))
                 item_factura_json['importeExento'] = float((datos_configuracion[0]['importe_exento']))
                 item_factura_json['importeNetoGravado'] = abs(float((sales_invoice_item[i]['facelec_amount_minus_excise_tax'])))
                 item_factura_json['importeOtrosImpuestos'] = float((sales_invoice_item[i]['facelec_other_tax_amount']))
@@ -332,7 +332,7 @@ def construir_xml(serie_original_factura, nombre_del_cliente, prefijo_serie, ser
             item_factura_json['cantidad'] = float(sales_invoice_item[0]['qty'])
             item_factura_json['codigoProducto'] = str(sales_invoice_item[0]['item_code'])
             item_factura_json['descripcionProducto'] = str((sales_invoice_item[0]['item_name']))
-            item_factura_json['detalleImpuestoIva'] = float((sales_invoice_item[0]['facelec_sales_tax_for_this_row']))
+            item_factura_json['detalleImpuestosIva'] = float((sales_invoice_item[0]['facelec_sales_tax_for_this_row']))
             item_factura_json['importeExento'] = float((datos_configuracion[0]['importe_exento']))
             item_factura_json['importeNetoGravado'] = abs(float((sales_invoice_item[0]['facelec_amount_minus_excise_tax'])))
             item_factura_json['importeOtrosImpuestos'] = float((sales_invoice_item[0]['facelec_other_tax_amount']))
